# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Beauty Analyzer API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  message: string;
  code?: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface RegisterRequest {
  name: string;
  email: string;
  password: string;
}

export interface UserProfile {
  id: string;
  name: string;
  email: string;
  avatar?: string;
  skinType?: string;
  age?: number;
  createdAt?: string;
}

export interface AuthResponse {
  token: string;
  user: UserProfile;
}

export interface UpdateProfileRequest {
  name?: string;
  skinType?: string;
  age?: number;
}

export interface CreateScanRequest {
  /** Base64 encoded image data */
  imageData: string;
  notes?: string;
}

export type ScanResultStatus =
  (typeof ScanResultStatus)[keyof typeof ScanResultStatus];

export const ScanResultStatus = {
  processing: "processing",
  completed: "completed",
  failed: "failed",
} as const;

export interface ScanResult {
  id: string;
  overallScore: number;
  createdAt: string;
  hydration?: number;
  oiliness?: number;
  acne?: number;
  darkSpots?: number;
  wrinkles?: number;
  sensitivity?: number;
  status: ScanResultStatus;
}

export type SkinMetricStatus =
  (typeof SkinMetricStatus)[keyof typeof SkinMetricStatus];

export const SkinMetricStatus = {
  good: "good",
  fair: "fair",
  poor: "poor",
} as const;

export interface SkinMetric {
  name: string;
  score: number;
  status: SkinMetricStatus;
  description?: string;
}

export interface Product {
  id: string;
  name: string;
  brand: string;
  category: string;
  price: number;
  rating: number;
  imageUrl?: string;
  description?: string;
  ingredients?: string[];
  isFavorite?: boolean;
}

export interface SkinReport {
  scanId: string;
  overallScore: number;
  skinType: string;
  concerns: string[];
  metrics: SkinMetric[];
  recommendations: string[];
  products?: Product[];
  createdAt?: string;
}

export type RoutineType = (typeof RoutineType)[keyof typeof RoutineType];

export const RoutineType = {
  morning: "morning",
  evening: "evening",
  weekly: "weekly",
} as const;

export interface RoutineStep {
  order: number;
  name: string;
  description?: string;
  product?: string;
  duration?: number;
  isCompleted?: boolean;
}

export interface Routine {
  id: string;
  name: string;
  type: RoutineType;
  steps: RoutineStep[];
  /** Duration in minutes */
  duration?: number;
  isCompleted?: boolean;
}

export type ImprovementTrend =
  (typeof ImprovementTrend)[keyof typeof ImprovementTrend];

export const ImprovementTrend = {
  improving: "improving",
  stable: "stable",
  declining: "declining",
} as const;

export interface Improvement {
  metric: string;
  change: number;
  trend: ImprovementTrend;
}

export interface ProgressData {
  scans: ScanResult[];
  improvements: Improvement[];
  streakDays: number;
  totalScans: number;
}