# lib/api-zod/src/generated/types/updateProfileRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Beauty Analyzer API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateProfileRequest {
  name?: string;
  skinType?: string;
  age?: number;
}