# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Beauty Analyzer API
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfile {
  id: string;
  name: string;
  email: string;
  avatar?: string;
  skinType?: string;
  age?: number;
  createdAt?: Date;
}